requirements.txt
adal==1.2.7
applicationinsights==0.11.10
argcomplete==3.5.1
asttokens==2.4.1
attrs==24.2.0
azure-common==1.1.28
azure-core==1.32.0
azure-graphrbac==0.61.2
azure-identity==1.19.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.0.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==21.2.1
azureml-automl-core==1.58.0
azureml-core==1.58.0
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.5
azureml-dataset-runtime==1.58.0
azureml-pipeline==1.58.0
azureml-pipeline-core==1.58.0
azureml-pipeline-steps==1.58.0
azureml-sdk==1.58.0
azureml-telemetry==1.58.0
azureml-train-automl-client==1.58.0
azureml-train-core==1.58.0
azureml-train-restclients-hyperdrive==1.58.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
cryptography==43.0.3
debugpy==1.8.9
decorator==5.1.1
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
fusepy==3.0.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.2.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.29.0
isodate==0.7.2
jedi==0.19.2
jeepney==0.8.0
jmespath==1.0.1
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
knack==0.12.0
matplotlib-inline==0.1.7
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
numpy==1.23.5
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
parso==0.8.4
pathspec==0.12.1
pkginfo==1.11.2
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
SecretStorage==3.3.3
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.21.0
